+++ cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xata.io/cli
 
+## 0.11.4
+
+### Patch Changes
+
+- [#808](https://github.com/xataio/client-ts/pull/808) [`463f740d`](https://github.com/xataio/client-ts/commit/463f740d8d272cbfd22330c9be0c31911edde5a2) Thanks [@SferaDev](https://github.com/SferaDev)! - Add schema upload command
+
+## 0.11.3
+
+### Patch Changes
+
+- [#804](https://github.com/xataio/client-ts/pull/804) [`96e49f35`](https://github.com/xataio/client-ts/commit/96e49f354edf0ff6a52047218b46c96056e82336) Thanks [@SferaDev](https://github.com/SferaDev)! - Fix race condition with workers build step
+
 ## 0.11.2
 
 ### Patch Changes